cleanup and error resolution attemp
general cleanup of some chart elements and adding some text to clarify some cells

Found at least a few instances where if the next iteration would produce a negative stress that if I injected instead a unit stress the solver would then find a solution so added an if statement to each vba routine to check if the next Y is <0 and if so instead use a value of 1. This is really a brute force attempt, ultimately need to find a better way to determine initial values to give the solver a better chance to find a solution.