cleanup and user notes
added some user notes and checks on the coordinate order for the plate and hole

changed the formatting for the results chart title to show the status label

changed the status label to indicate design status and solver status rather than just "NG and Solution Found" 

Started work on a solver to find the minimum square area projection from the load point that results in all compressive stress and peak stress less than Fp,max. This is will end up equivalent to how RAM Base Plate would allow the user the option to only design to the required bearing area instead of using all of the available plate area.