additional initial guess point
added one more guess point to the grid to check against with a stress value = 10x the elastic stress, so far this has reached solutions for almost all the cases that were diverging. There are a few outliers where a slightly relaxed tolerance results in a solution, not sure if this may be an excel rounding error since it's down near 1E-12.