working out triangular plates
so far for some configurations of moment only the solver will fail with triangular or trapezoids, where the smallest side is less than 4 in. If a very small axial load, P=1E-5, is added the solver then is able to find a solution. Working on various tests cases to try and track down why this may be occurring.