found an OK starting point for triangles/trapezoids
Found an additional guess point that leads to convergence with triangles/trapezoids. The point works as is now but can be improved per the below.

With triangles I found that in some instances one of the vertices has 0 stress using an elastic analysis which fell outside of the logic I was using to determine the elastic neutral axis, fixing this and using that neutral axis angle with the new guess point cuts the number of iterations in half. This isn't implemented yet, it's a little tricky.