Secant Method to Find Neutral Axis Depth
added in a macro that uses the Secant Method to find the neutral axis depth with a fall back to the bisection method if needed. Bisection method was taking 45-50 iterations to find the right neutral axis depth the Secant method is finding solutions within 7 iterations which significantly decreases the run time for the neutral axis angle solvers that rely on intermediate calculations of the neutral axis depth.